:sparkles: [op] Add and use WireAPI, RPCServer and RPCOperator.
Add the RPCServer struct which serves WS connections and parses
messages. It then forwards the calls to the RPCOperator who
handles proof subscriptions and adds transactions.
The Operator starts and maintains the RPCOperator, such that
the client is always served the latest proofs.

The WireAPI interface is needed for mocking only.

Signed-off-by: Oliver Tale-Yazdi <oliver@perun.network>