:wrench: [op] Fix derivation path and logging.
Having the Operator on account index 0 is preferable since it will
do on-chain transactions and wallets like metamask can only access
the wallet with index 0. It is easier to test and deploy this way.
Add the Operators address to the accounts.json so its on-chain
balance can be monitored.
Log the Operator and Enclave addresses for easier deploying.

Signed-off-by: Oliver Tale-Yazdi <oliver@perun.network>