contract: Make sure only active system participants can challenge.
The contract had two bugs:

1. Anyone could post a challenge even if they never entered the system.
This would effectively freeze the contract because the operator couldn't
answer a challenge by an unknown user.
2. Once the contract was frozen, a user could withdraw with any balance
proof, not just the frozen epoch's one.

Both bugs are now mitigated by, when posting a challenge, having the
user either post the balance proof of the last sealed epoch, proofing
that they are part of the system; or it is implicitly required that the
user made a deposit in the challenged epoch. This also means that the
to be recovered value is already determined and stored when challenging.