[operator] Added ClientPush logic
Before clients had to know the address of the Erdstall-Contract as well
as chainID to connect with the operator and be able to join the Erdstall
System.
This change allows the operator to push a `ClientConfig` over the wire
containing ChainID/NetworkID as well as Contract-Address. This allows a
client to set up all web3 related connections upon connection to theh
operator.

Signed-off-by: Norbert Dzikowski <norbert@perun.network>